add test cases / refactor the autocomplete code to extract the actions.
