+++ hamster-selenium-component-mat/src/main/java/com/github/grossopa/hamster/selenium/component/mat/action/AutocompleteCloseOptionsAction.java
@@ -22,59 +22,27 @@
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.github.grossopa.hamster.selenium.component.mat.config;
+package com.github.grossopa.hamster.selenium.component.mat.action;
+
+import com.github.grossopa.hamster.selenium.component.mat.main.MatAutocomplete;
+import com.github.grossopa.selenium.core.ComponentWebDriver;
+import com.github.grossopa.selenium.core.component.WebComponent;
+
+import java.util.List;
+
+import static org.openqa.selenium.Keys.ESCAPE;
 
 /**
- * The root configuration for Material UI Angular components
+ * For {@link MatAutocomplete} to close the options by sending a {@link org.openqa.selenium.Keys#ESCAPE} key event to
+ * the autocomplete input element.
  *
  * @author Jack Yin
  * @since 1.6
  */
-public class MatConfig {
+public class AutocompleteCloseOptionsAction implements CloseOptionsAction {
 
-    /**
-     * Attribute class
-     */
-    public static final String ATTR_CLASS = "class";
-
-    private String tagPrefix = "mat-";
-
-    private String cssPrefix = "mat-";
-
-    private String cdkPrefix = "cdk-";
-
-    @SuppressWarnings("java:S1075")
-    private String overlayAbsolutePath = "/html/body";
-
-    public String getTagPrefix() {
-        return tagPrefix;
-    }
-
-    public void setTagPrefix(String tagPrefix) {
-        this.tagPrefix = tagPrefix;
-    }
-
-    public String getCssPrefix() {
-        return cssPrefix;
-    }
-
-    public void setCssPrefix(String cssPrefix) {
-        this.cssPrefix = cssPrefix;
-    }
-
-    public String getOverlayAbsolutePath() {
-        return overlayAbsolutePath;
-    }
-
-    public void setOverlayAbsolutePath(String overlayAbsolutePath) {
-        this.overlayAbsolutePath = overlayAbsolutePath;
-    }
-
-    public String getCdkPrefix() {
-        return cdkPrefix;
-    }
-
-    public void setCdkPrefix(String cdkPrefix) {
-        this.cdkPrefix = cdkPrefix;
+    @Override
+    public void close(WebComponent ownerComponent, List<? extends WebComponent> options, ComponentWebDriver driver) {
+        ((MatAutocomplete) ownerComponent).getInput().sendKeys(ESCAPE);
     }
 }
